# resources/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "rewear.id API",
    "description": "API documentation for rewear.id — Marketplace Jual Beli Pakaian Preloved dengan Sistem Escrow.\n\n## Authentication\nAll authenticated endpoints require a Bearer token from Sanctum. Include it in the `Authorization` header:\n```\nAuthorization: Bearer {token}\n```\n\n## Flow Utama\n1. **Register/Login** → Dapatkan token\n2. **Browse Products** → Cari pakaian preloved\n3. **Create Order** → Buat pesanan\n4. **Payment** → Bayar via Midtrans\n5. **Escrow** → Dana ditahan sampai barang diterima\n6. **Confirm Received** → Pembeli konfirmasi → Dana dilepas ke penjual",
    "version": "1.0.0",
    "contact": {
      "name": "rewear.id",
      "email": "support@rewear.id"
    }
  },
  "servers": [
    {
      "url": "{protocol}://{host}",
      "variables": {
        "protocol": { "default": "http", "enum": ["http", "https"] },
        "host": { "default": "127.0.0.1:8000" }
      }
    }
  ],
...
  "paths": {
...
    "/api/v1/addresses/{address}": {
      "put": {
        "tags": ["Addresses"], "summary": "Update address", "operationId": "addresses.update",
        "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "address", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddressRequest" } } } },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Addresses"], "summary": "Delete address", "operationId": "addresses.destroy",
        "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "address", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } }
      }
    },
...
    "/api/v1/bank-accounts/{bankAccount}": {
      "delete": {
        "tags": ["Bank Accounts"], "summary": "Delete bank account", "operationId": "bankAccounts.destroy",
        "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "bankAccount", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } }
      }
    },
...
    "/api/v1/seller/products/{product}": {
      "put": {
        "tags": ["Seller Products"], "summary": "Update product", "operationId": "seller.products.update",
        "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "product", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductRequest" } } } },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Seller Products"], "summary": "Delete product", "operationId": "seller.products.destroy",
        "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "product", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } }
      }
    },
...
    "/api/v1/seller/products/{product}/images/{image}": {
      "delete": {
        "tags": ["Seller Products"], "summary": "Delete product image", "operationId": "seller.products.images.destroy",
        "security": [{ "sanctum": [] }],
        "parameters": [
          { "name": "product", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "image", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } }
      }
    },
...
    "/api/v1/bookmarks/{product}": {
      "delete": {
        "tags": ["Bookmarks"], "summary": "Remove bookmark", "operationId": "bookmarks.destroy",
        "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "product", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } }
      }
    },
...
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "tags": ["Notifications"], "summary": "Mark notification as read", "operationId": "notifications.markAsRead",
        "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } }
      }
    },
    "/api/v1/notifications/read-all": {
      "patch": {
        "tags": ["Notifications"], "summary": "Mark all notifications as read", "operationId": "notifications.markAllAsRead",
        "security": [{ "sanctum": [] }],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } }
      }
    },
...
    "/api/v1/admin/products/{productId}": {
      "delete": { "tags": ["Admin - Products"], "summary": "Remove product (moderation)", "operationId": "admin.products.destroy", "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "productId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "$ref": "#/components/responses/MessageResponse" } } }
    },
...
    "/api/v1/admin/withdrawals/{withdrawal}/approve": {
      "post": { "tags": ["Admin - Withdrawals"], "summary": "Approve withdrawal", "operationId": "admin.withdrawals.approve", "security": [{ "sanctum": [] }],
        "parameters": [{ "name": "withdrawal", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Withdrawal approved" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum Token",
        "description": "Enter the token from login/register response"
      }
    },
    "schemas": {
...
      "AddressRequest": {
        "type": "object",
        "required": ["label", "recipient_name", "phone", "province_id", "city_id", "district", "postal_code", "full_address"],
        "properties": {
          "label": { "type": "string", "maxLength": 50, "example": "Rumah" },
          "recipient_name": { "type": "string", "maxLength": 100, "example": "John Doe" },
          "phone": { "type": "string", "maxLength": 20, "example": "08123456789" },
          "province_id": { "type": "integer", "example": 6 },
          "city_id": { "type": "integer", "example": 152 },
          "district": { "type": "string", "maxLength": 100, "example": "Kebayoran Baru" },
          "postal_code": { "type": "string", "maxLength": 10, "example": "12160" },
          "full_address": { "type": "string", "example": "Jl. Sudirman No. 1" },
          "is_default": { "type": "boolean" }
        }
      },
...
      "ProductRequest": {
        "type": "object",
        "required": ["category_id", "title", "description", "size", "condition", "price", "weight_grams"],
        "properties": {
          "category_id": { "type": "string" },
          "brand_id": { "type": "string", "nullable": true },
          "title": { "type": "string", "maxLength": 200, "example": "Kemeja Flannel Uniqlo Size L" },
          "description": { "type": "string", "example": "Kondisi like new, baru dipakai 2x" },
          "size": { "type": "string", "maxLength": 10, "example": "L" },
          "condition": { "type": "string", "enum": ["new_with_tag", "like_new", "good", "fair"] },
          "color": { "type": "string", "maxLength": 50, "nullable": true, "example": "Navy" },
          "price": { "type": "number", "minimum": 1000, "example": 150000 },
          "weight_grams": { "type": "integer", "minimum": 1, "example": 300 },
          "images[]": { "type": "array", "items": { "type": "string", "format": "binary" } }
        }
      },
...
    },
    "responses": {
      "MessageResponse": {
        "description": "Success message",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" } } } } }
      },
...
    }
  }
}